+++ gleam_stdlib/src/gleam_stdlib.erl
@@ -3,7 +3,7 @@
 
 -export([expect_equal/2, expect_not_equal/2, expect_true/1, expect_false/1,
          expect_is_ok/1, expect_is_error/1, atom_from_string/1,
-         atom_create_from_string/1, atom_to_string/1, map_fetch/2,
+         atom_create_from_string/1, atom_to_string/1, map_get/2,
          iodata_append/2, iodata_prepend/2, identity/1, decode_int/1,
          decode_string/1, decode_bool/1, decode_float/1, decode_thunk/1, decode_atom/1,
          decode_pair/1, decode_list/1, decode_field/2, parse_int/1, parse_float/1, compare_strings/2]).
@@ -15,7 +15,7 @@ expect_false(A) -> ?assertNot(A).
 expect_is_ok(A) -> ?assertMatch({ok, _}, A).
 expect_is_error(A) -> ?assertMatch({error, _}, A).
 
-map_fetch(Map, Key) ->
+map_get(Map, Key) ->
   case maps:find(Key, Map) of
     error -> {error, nil};
     OkFound -> OkFound