+++ gleam_stdlib/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- `map.fetch` has been renamed to `map.get` and `map.put` to `map.insert`.
 - The `pair` module gains the `map_first`, and `map_second` functions.
 - The `string` module gains the `compare` function.
 - The `float` module gains the `max`,  and `min` functions.